restore old tests, to try and make a clean commit, again

+++ tests/v2/test_storage.py
@@ -2221,45 +2221,44 @@ def test_format_compatibility():
     # read data stored with a previous minor version (which should be format-compatible).
 
     # fixture data
-    fixture_path = pathlib.Path(__file__).parent / "fixture" / "test_format_compatibility"
-    fixture = group(store=DirectoryStore(fixture_path))
+    fixture = group(store=DirectoryStore("fixture"))
 
     # set seed to get consistent random data
     np.random.seed(42)
 
     arrays_chunks = [
-        (np.arange(1111, dtype="<i1"), 600),
-        (np.arange(1111, dtype="<i2"), 600),
-        (np.arange(1111, dtype="<i4"), 600),
+        (np.arange(1111, dtype="<i1"), 100),
+        (np.arange(1111, dtype="<i2"), 100),
+        (np.arange(1111, dtype="<i4"), 100),
         (np.arange(1111, dtype="<i8"), 1000),
-        (np.random.randint(0, 200, size=2222, dtype="u1").astype("<u1"), 1200),
-        (np.random.randint(0, 2000, size=2222, dtype="u2").astype("<u2"), 1200),
-        (np.random.randint(0, 2000, size=2222, dtype="u4").astype("<u4"), 1200),
-        (np.random.randint(0, 2000, size=2222, dtype="u8").astype("<u8"), 1200),
-        (np.linspace(0, 1, 3333, dtype="<f2"), 2000),
-        (np.linspace(0, 1, 3333, dtype="<f4"), 2000),
-        (np.linspace(0, 1, 3333, dtype="<f8"), 2000),
-        (np.random.normal(loc=0, scale=1, size=4444).astype("<f2"), 3000),
-        (np.random.normal(loc=0, scale=1, size=4444).astype("<f4"), 3000),
-        (np.random.normal(loc=0, scale=1, size=4444).astype("<f8"), 3000),
-        (np.random.choice([b"A", b"C", b"G", b"T"], size=5555, replace=True).astype("S"), 3000),
+        (np.random.randint(0, 200, size=2222, dtype="u1").astype("<u1"), 100),
+        (np.random.randint(0, 2000, size=2222, dtype="u2").astype("<u2"), 100),
+        (np.random.randint(0, 2000, size=2222, dtype="u4").astype("<u4"), 100),
+        (np.random.randint(0, 2000, size=2222, dtype="u8").astype("<u8"), 100),
+        (np.linspace(0, 1, 3333, dtype="<f2"), 100),
+        (np.linspace(0, 1, 3333, dtype="<f4"), 100),
+        (np.linspace(0, 1, 3333, dtype="<f8"), 100),
+        (np.random.normal(loc=0, scale=1, size=4444).astype("<f2"), 100),
+        (np.random.normal(loc=0, scale=1, size=4444).astype("<f4"), 100),
+        (np.random.normal(loc=0, scale=1, size=4444).astype("<f8"), 100),
+        (np.random.choice([b"A", b"C", b"G", b"T"], size=5555, replace=True).astype("S"), 100),
         (
             np.random.choice(["foo", "bar", "baz", "quux"], size=5555, replace=True).astype("<U"),
-            3000,
+            100,
         ),
         (
             np.random.choice([0, 1 / 3, 1 / 7, 1 / 9, np.nan], size=5555, replace=True).astype(
                 "<f8"
             ),
-            3000,
+            100,
         ),
-        (np.random.randint(0, 2, size=5555, dtype=bool), 3000),
-        (np.arange(20000, dtype="<i4").reshape(2000, 10, order="C"), (2000, 5)),
-        (np.arange(20000, dtype="<i4").reshape(200, 100, order="F"), (200, 50)),
-        (np.arange(20000, dtype="<i4").reshape(200, 10, 10, order="C"), (200, 10, 5)),
-        (np.arange(20000, dtype="<i4").reshape(20, 100, 10, order="F"), (20, 100, 10)),
-        (np.arange(20000, dtype="<i4").reshape(20, 10, 10, 10, order="C"), (20, 10, 10, 5)),
-        (np.arange(20000, dtype="<i4").reshape(20, 10, 10, 10, order="F"), (20, 10, 10, 5)),
+        (np.random.randint(0, 2, size=5555, dtype=bool), 100),
+        (np.arange(20000, dtype="<i4").reshape(2000, 10, order="C"), (100, 3)),
+        (np.arange(20000, dtype="<i4").reshape(200, 100, order="F"), (100, 30)),
+        (np.arange(20000, dtype="<i4").reshape(200, 10, 10, order="C"), (100, 3, 3)),
+        (np.arange(20000, dtype="<i4").reshape(20, 100, 10, order="F"), (10, 30, 3)),
+        (np.arange(20000, dtype="<i4").reshape(20, 10, 10, 10, order="C"), (10, 3, 3, 3)),
+        (np.arange(20000, dtype="<i4").reshape(20, 10, 10, 10, order="F"), (10, 3, 3, 3)),
     ]
 
     compressors = [
@@ -2279,14 +2278,11 @@ def test_format_compatibility():
             order = "C"
 
         for j, compressor in enumerate(compressors):
-            path = f"array_{i}/compressor_{j}"
+            path = "{}/{}".format(i, j)
 
-            # uncomment these lines to generate the fixture files, which should be unnecessary 
-            # since they are checked into version control
-            
-            # if path not in fixture:  # pragma: no cover
-            #     # store the data - should be one-time operation
-            #     fixture.array(path, data=arr, chunks=chunks, order=order, compressor=compressor)
+            if path not in fixture:  # pragma: no cover
+                # store the data - should be one-time operation
+                fixture.array(path, data=arr, chunks=chunks, order=order, compressor=compressor)
 
             # setup array
             z = fixture[path]
@@ -2460,17 +2456,14 @@ def test_normalize_store_arg(tmpdir):
         store = normalize_store_arg(fsspec.get_mapper("file://" + path))
         assert isinstance(store, FSStore)
 
-# todo: delete this
-def test_meta_prefix_6853(tmpdir):
-    """
-    This test ensures that groups can start with the name "meta". Added to resolve 
-    the issue described here: https://github.com/zarr-developers/zarr-python/issues/1105
-    """
-    fixture = pathlib.Path(str(tmpdir)) /  "fixture"
+
+def test_meta_prefix_6853():
+    fixture = pathlib.Path(zarr.v2.__file__).resolve().parent.parent / "fixture"
     meta = fixture / "meta"
-    s = DirectoryStore(str(meta), dimension_separator=".")
-    a = zarr.v2.open(store=s, mode="w", shape=(2, 2), dtype="<i8")
-    a[:] = [[1, 2], [3, 4]]
+    if not meta.exists():  # pragma: no cover
+        s = DirectoryStore(str(meta), dimension_separator=".")
+        a = zarr.v2.open(store=s, mode="w", shape=(2, 2), dtype="<i8")
+        a[:] = [[1, 2], [3, 4]]
 
     fixtures = group(store=DirectoryStore(str(fixture)))
     assert list(fixtures.arrays())
